requirements.txt
# Core
python-telegram-bot==21.6
anthropic>=0.28.0
httpx>=0.27.0
python-dotenv>=1.0.0

# PDF parsing
pdfplumber>=0.11.0
pypdf>=4.0.0

# Web scraping (Djinni, LinkedIn, Glassdoor)
beautifulsoup4>=4.12.0
playwright>=1.44.0          # run: playwright install chromium

# Optional: Telethon for Telegram channel scanning with user account
# telethon>=1.36.0
